api/t/net: move endpoint structs into endpoint.go
